feat: partial upgrade Apache PDFBox from 2.0.34 to 3.0.5
- Updated PDFBox version to 3.0.5 in build.gradle
- Updated font constants to use new PDType1Font(Standard14Fonts.FontName.HELVETICA) syntax
- Updated PDDocument.load() calls to use new Loader.loadPDF() API with byte arrays
- Updated drawing methods from drawLine(), fillRect() to new path construction API
- Updated signature field API calls by removing getWidget() method calls
- Updated Matrix API calls from setFromAffineTransform() to constructor-based approach
- Updated COSNumber API usage from doubleValue() to floatValue()
- Partially updated OperatorProcessor constructor usage

Note: This is a partial migration due to extensive API changes in PDFBox 3.x.
Additional work is needed to complete the migration including:
- OperatorProcessor architecture refactoring
- Signature field method updates
- PreflightParser API updates
- XMP metadata API updates
- Various other API method changes

+++ pdf-as-pdfbox-2/src/main/java/at/knowcenter/wag/egov/egiz/pdfbox2/pdf/operator/path/construction/CurveToReplicateInitialPoint.java
@@ -86,8 +86,8 @@ public class CurveToReplicateInitialPoint extends PathConstructionOperatorProces
 			COSNumber y3 = (COSNumber) operands.get(3);
 
 	        Point2D currentPoint = pdfPage.getCurrentPath().getCurrentPoint();
-			Point2D p2 = transform(x2.doubleValue(), y2.doubleValue());
-			Point2D p3 = transform(x3.doubleValue(), y3.doubleValue());
+			Point2D p2 = transform(x2.floatValue(), y2.floatValue());
+			Point2D p3 = transform(x3.floatValue(), y3.floatValue());
 
 			pdfPage.getCurrentPath().curveTo(
 					(float)currentPoint.getX(), (float)currentPoint.getY(),
